fix(ci): drop redundant casts in insurance_pool validation
`contracts/insurance_pool/src/lib.rs` had two `as u32` casts on
`.len()` calls in claim-validation guards (lines 571, 576). Soroban
`Vec::len()` and `String::len()` already return `u32`, so the cast
was redundant. Drops both and lets clippy through under
`-D unnecessary_cast`.

These two cases were the actual blockers on the GitHub Actions
`clippy` workflow run for `9e2476b`; the local capture had been
truncated before this crate was reached.

+++ contracts/insurance_pool/src/lib.rs
@@ -568,12 +568,12 @@ impl InsurancePoolContract {
         }
 
         // Validate description length
-        if description.len() as u32 > MAX_DESCRIPTION_LENGTH {
+        if description.len() > MAX_DESCRIPTION_LENGTH {
             panic!("Description too long");
         }
 
         // Validate evidence hash length
-        if evidence_hash.len() as u32 > MAX_EVIDENCE_HASH_LENGTH {
+        if evidence_hash.len() > MAX_EVIDENCE_HASH_LENGTH {
             panic!("Evidence hash too long");
         }
